Moves packages and creates steps file.
