+++ app/src/main/java/udacity/lsaippa/cooking/ui/detail/recipe/DetailActivity.java
@@ -1,4 +1,4 @@
-package udacity.lsaippa.cooking.ui.detail;
+package udacity.lsaippa.cooking.ui.detail.recipe;
 
 import android.os.Bundle;
 import android.support.annotation.Nullable;
@@ -12,8 +12,6 @@ import android.widget.Toast;
 import udacity.lsaippa.cooking.R;
 import udacity.lsaippa.cooking.network.model.Recipe;
 import udacity.lsaippa.cooking.network.model.Step;
-import udacity.lsaippa.cooking.ui.detail.recipe.DetailRecipeAdapter;
-import udacity.lsaippa.cooking.ui.detail.recipe.DetailRecipeFragment;
 
 import static udacity.lsaippa.cooking.utils.AppConstants.RECIPE_TAG;
 
@@ -23,7 +21,7 @@ public class DetailActivity extends AppCompatActivity implements DetailRecipeAda
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_detail);
+        setContentView(R.layout.activity_detail_recipe);
 
         Bundle bundle = getIntent().getExtras();
         if (bundle != null || getIntent().hasExtra(RECIPE_TAG)) {
