+++ app/src/main/java/udacity/lsaippa/cooking/ui/detail/step/DetailStepActivity.java
@@ -1,0 +1,16 @@
+package udacity.lsaippa.cooking.ui.detail.step;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v7.app.AppCompatActivity;
+
+import udacity.lsaippa.cooking.R;
+
+public class DetailStepActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_detail_step);
+    }
+}
