+++ app/src/main/java/udacity/lsaippa/cooking/ui/detail/step/DetailStepAdapter.java
@@ -1,0 +1,5 @@
+package udacity.lsaippa.cooking.ui.detail.step;
+
+
+public class DetailStepAdapter {
+}
