+++ app/src/main/java/udacity/lsaippa/cooking/ui/detail/step/DetailStepFragment.java
@@ -1,0 +1,6 @@
+package udacity.lsaippa.cooking.ui.detail.step;
+
+import android.support.v4.app.Fragment;
+
+public class DetailStepFragment extends Fragment {
+}
