--- app/src/main/java/udacity/lsaippa/cooking/ui/detail/step/StepFragment.java
@@ -1,8 +1,0 @@
-package udacity.lsaippa.cooking.ui.detail.step;
-
-/**
- * Created by lsaippa on 12/01/19.
- */
-
-public class StepFragment {
-}
