+++ app/src/main/res/layout/activity_detail_recipe.xml
